uct-benchmark-pipeline/requirements.txt
# UCT Benchmark Pipeline Dependencies
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
python-dateutil>=2.8.2
